project c 0x14

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  */
 void print_binary(unsigned long int n)
 {
-unsigned int l = 0, max = 32768;
+unsigned long int l = 0, max = 32768;
 if (n == 0)
 {
 _putchar('0');
